ai/requirements.txt
fastapi
openai
langchain
pinecone-client
sentence-transformers
tiktoken
motor
websockets
pydantic